+++ lizard_datasource/tests/test_timeseries.py
@@ -1,6 +1,9 @@
 """Tests for lizard_datasource.timeseries"""
 
+import StringIO
+
 import pandas
+import pytz
 
 from django.test import TestCase
 
@@ -44,3 +47,25 @@ class TestTimeseries(TestCase):
     def test_timeseries_has_a_length(self):
         ts = timeseries.Timeseries({self.some_date: self.some_value})
         self.assertEquals(len(ts), 1)
+
+    def test_two_timeseries_to_csv(self):
+        dt1 = dates.utc(2014, 6, 24, 17, 31, 24)
+        dt2 = dates.utc(2014, 6, 24, 17, 32, 24)
+        dt3 = dates.utc(2014, 6, 24, 17, 33, 24)
+
+        ts = timeseries.Timeseries([
+            {dt1: 1.0, dt2: 2.0},
+            {dt1: 3.0, dt3: 4.0}
+            ])
+
+        outfile = StringIO.StringIO()
+
+        ts.to_csv(outfile, timezone=pytz.timezone('Europe/Amsterdam'))
+
+        # Summer, so two hours difference with UTC
+        self.assertEquals(outfile.getvalue(),
+                          """Datum + tijd,data_0,data_1
+2014-06-24 19:31,1.0,3.0
+2014-06-24 19:32,2.0,
+2014-06-24 19:33,,4.0
+""")